graph comparison func, some tests cleanup
