+++ tests/parser/tests_JSON_LD_Parser.cpp
@@ -2,54 +2,21 @@
 #include <doctest/doctest.h>
 #include <rdf4cpp.hpp>
 
-#include <filesystem>
-#include <curl/curl.h>
-
 #include "parser_test_helpers.hpp"
 
 using namespace rdf4cpp;
 using namespace rdf4cpp::parser;
 
 void jsonld_test_positive(std::string json_str, std::string nt_str, std::string_view base_iri, bool dedup = false, ParsingFlag dir = ParsingFlag::JsonLdDirectionNone) {
-    parse_test_helpers::jsonld_test_positive(std::move(json_str), std::move(nt_str), base_iri, ParsingFlag::JsonLd | dir, ParsingFlag::NQuads, dedup);
+    parse_test_helpers::parser_test_positive(std::move(json_str), std::move(nt_str), base_iri, ParsingFlag::JsonLd | dir, ParsingFlag::NQuads, dedup);
 }
 
 void jsonld_test_negative(std::string json_str, std::string_view base_iri) {
     parse_test_helpers::parser_test_negative(std::move(json_str), base_iri, ParsingFlag::JsonLd);
 }
 
-// adopted from https://stackoverflow.com/questions/9786150/save-curl-content-result-into-a-string-in-c/9786295#9786295
-static size_t write_callback(void const *contents, size_t size, size_t nmemb, void *userp) {
-    static_cast<std::string *>(userp)->append(static_cast<char const *>(contents), size * nmemb);
-    return size * nmemb;
-}
-
-std::string remote_test_file_to_str(std::string const &file_name) {
-    if (!std::filesystem::exists("./jsonld_test_cache")) { // this ends up in the cmake folder, next to the executable
-        std::filesystem::create_directories("./jsonld_test_cache");
-    }
-    auto cache = std::format("./jsonld_test_cache/{}", file_name);
-    if (std::filesystem::exists(cache)) {
-        std::ifstream ifs(cache);
-        return std::string{std::istreambuf_iterator{ifs}, {}};
-    }
-    CURL *curl = nullptr;
-    CURLcode curl_res;
-    auto const url = std::format("https://raw.githubusercontent.com/w3c/json-ld-streaming/refs/heads/main/tests/stream-toRdf/{}", file_name);
-    std::string file_contents_as_str;
-    curl = curl_easy_init();
-    if(curl) {
-        curl_easy_setopt(curl, CURLOPT_URL, url.c_str());
-        curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, write_callback);
-        curl_easy_setopt(curl, CURLOPT_WRITEDATA, &file_contents_as_str);
-        curl_easy_setopt(curl, CURLOPT_SSL_VERIFYPEER, 0L); // for https
-        curl_res = curl_easy_perform(curl);
-        curl_easy_cleanup(curl);
-    }
-    REQUIRE_EQ(curl_res, CURLE_OK);
-    std::ofstream ofs(cache);
-    ofs << file_contents_as_str;
-    return file_contents_as_str;
+std::string remote_test_file_to_str(std::string_view file_name) {
+    return parse_test_helpers::parser_test_remote_test_file_to_str(file_name, "https://raw.githubusercontent.com/w3c/json-ld-streaming/refs/heads/main/tests/stream-toRdf", "./jsonld_test_cache");
 }
 
 TEST_CASE("test cases from rdf-tests") {
@@ -89,7 +56,7 @@ TEST_CASE("test cases from rdf-tests") {
     jsonld_test_positive(remote_test_file_to_str("0034-in.jsonld"), remote_test_file_to_str("0034-out.nq"), "https://w3c.github.io/json-ld-streaming/tests/t0034");
     // 9.9 as int
     //jsonld_test_positive(remote_test_file_to_str("0035-in.jsonld"), remote_test_file_to_str("0035-out.nq"), "https://w3c.github.io/json-ld-streaming/tests/t0035");
-    parse_test_helpers::jsonld_test_positive(R"({
+    parse_test_helpers::parser_test_positive(R"({
   "@context": {
     "double": {
       "@id": "http://example.com/double",
